fastapi2/requirements-frozen.txt
absl-py==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
arxiv==2.4.0
attrs==25.4.0
bcrypt==5.0.0
brotli==1.2.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
coverage==7.13.2
cryptography==46.0.4
cssselect2==0.8.0
cycler==0.12.1
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
ecdsa==0.19.1
fastapi==0.128.0
feedparser==6.0.12
filelock==3.20.3
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.1.0
google-auth==2.48.0
google-genai==1.61.0
greenlet==3.3.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ImageIO==2.37.2
iniconfig==2.3.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.2.8
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.13
langchain-google-genai==4.2.0
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
lazy_loader==0.4
llama_cpp_python==0.3.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.18
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
networkx==3.6.1
numpy==1.26.4
onnxruntime==1.24.1
opencv-contrib-python==4.11.0.86
opencv-python==4.9.0.80
opt_einsum==3.4.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.0
passlib==1.7.4
pillow==12.1.0
pip-tools==7.5.3
playwright==1.58.0
pluggy==1.6.0
propcache==0.4.1
protobuf==4.25.8
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.0
Pygments==2.19.2
pyparsing==3.3.2
pyphen==0.17.2
pyproject_hooks==1.2.0
pyserial==3.5
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
reportlab==4.4.9
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.2
rsa==4.9.1
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
sentencepiece==0.2.1
setuptools==80.10.2
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
SQLAlchemy==2.0.46
starlette==0.50.0
sympy==1.14.0
tavily-python==0.7.21
tenacity==9.1.4
threadpoolctl==3.6.0
tifffile==2026.1.28
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.2
transformers==5.0.0
typer-slim==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
watchfiles==1.1.1
weasyprint==68.0
webencodings==0.5.1
websockets==15.0.1
wheel==0.46.3
xmltodict==1.0.3
xxhash==3.6.0
yarl==1.22.0
zopfli==0.4.0
zstandard==0.25.0
